requirements.txt
# МособлГосЭкспертиза — DocumentAnalyzer v2.0
# Установка: pip install -r requirements.txt
# Chromium (Playwright): playwright install chromium

# ─── Telegram Bot ────────────────────────────
python-telegram-bot==22.6
python-dotenv>=1.0.0

# ─── LLM (Groq API) ──────────────────────────
groq>=1.0.0

# ─── HTTP / API ──────────────────────────────
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0

# ─── Парсинг XML / PDF / OCR ──────────────────
lxml>=5.0.0
xmlschema>=3.0.0
pymupdf>=1.24.0          # fitz — извлечение текста из PDF
pytesseract>=0.3.10      # OCR для извлечения текста из сканов
pdf2image>=1.17.0        # Конвертация страниц PDF-сканов в изображения

# ─── База данных (HITL / DisagreementLog) ─────
sqlalchemy>=2.0.0

# ─── Векторная БД (RAG/Weaviate) ─────────────
weaviate-client>=3.24.0,<4.0.0

# ─── Веб-скрейпинг (НОПРИЗ) ──────────────────
playwright>=1.40.0
# После установки: playwright install chromium

# ─── Pydantic ────────────────────────────────
pydantic>=2.0.0

# ─── PDF-отчёты (ГОСТ Р 7.0.97-2016) ─────────
reportlab>=4.0.0

# ─── Тесты ───────────────────────────────────
pytest>=7.0.0
pytest-asyncio>=0.21.0
